+++ clinicadl/data/structures/__init__.py
@@ -1,0 +1,2 @@
+from .datapoint import DataPoint
+from .label import Column, Mask
